Don't cast data/durability values to bytes because they're not always going to fit. Some items use values that are too high and casting causes us to lose information. Fixes a regression of bug 242 - bug 483:resolved

+++ src/main/java/me/botsko/prism/MaterialAliases.java
@@ -46,7 +46,7 @@ public class MaterialAliases {
 	 * @param subid
 	 * @return
 	 */
-	public String getItemStackAliasById( int typeid, byte subid ){
+	public String getItemStackAliasById( int typeid, int subid ){
 		String item_name = null;
 		if(!itemAliases.isEmpty()){
 			String key = typeid+":"+subid;
